+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 - Hide/Flatten non-searched buildings #225
 - Hide/Flatten all buildings, if searchPattern can't find any matching nodes #225
 - Colored color-slider inside the RibbonBar #318
+- Color positive buildings white and set as default option #311
 
 ### Changed
 - Update TreeView filter with search field #225